initial 2

+++ app/src/main/java/com/example/profitportfolio/DbHelper.java
@@ -9,22 +9,24 @@ import androidx.annotation.Nullable;
 public class DbHelper extends SQLiteOpenHelper {
     public static final String DBNAME="stock.db";
     public static final String TABLENAME="stock";
-    public static final int VER=1;
+    public static final int VER=2;
     public DbHelper(@Nullable Context context) {
         super(context, TABLENAME, null, VER);
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        String sql = "create table " + TABLENAME + "(id integer primary key, name text, pieces text, buyDate text, sellDate text, stockPriceBuy REAL, stockPriceSell REAL, amount REAL, profitAndLoss REAL, komisyon REAL)";
+        String sql = "create table " + TABLENAME + "(id integer primary key, name text, pieces text, buyDate text, sellDate text, stockPriceBuy REAL, stockPriceSell REAL, amount REAL, profitAndLoss REAL, komisyon REAL,total REAL)";
         db.execSQL(sql);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        String sql = "drop table if exists " + TABLENAME + " ";
-        db.execSQL(sql);
-        onCreate(db);
+        if (oldVersion < newVersion) {
+            String sql = "drop table if exists " + TABLENAME + " ";
+            db.execSQL(sql);
+            onCreate(db);
+        }
     }
 
 }
